refactor: Extract VSCode extension logic to its own module
